+++ fe1-web/features/dependencies.md
@@ -12,3 +12,10 @@ the future.
 - SocialHandler : `ingestion/handlers/index.ts`
 - SocialNavigation : `navigation/bars/LaoNavigation.tsx`,
 - SocialReducer : `store/reducers/index.ts`, `store/reducers/RootReducer.ts`
+
+## Meeting
+
+- EventMeeting : `components/eventList/events/Event.tsx`
+- Meeting : `model/objects/LaoEventBuilder.ts`, `model/objects/__tests__/LaoEventBuilder.test.ts`
+- MeetingHandler : `ingestion/handlers/index.ts`
+- Storage is done in EventsReducer/EventsStore and could be done separately in a MeetingsReducer
