+++ fe1-web/features/meeting/network/messages/CreateMeeting.ts
@@ -4,8 +4,8 @@ import {
 import { OpenedLaoStore } from 'store';
 import { ProtocolError } from 'model/network/ProtocolError';
 import { validateDataObject } from 'model/network/validation';
-import { ActionType, MessageData, ObjectType } from '../MessageData';
-import { checkTimestampStaleness } from '../Checker';
+import { ActionType, MessageData, ObjectType } from 'model/network/method/message/data/MessageData';
+import { checkTimestampStaleness } from 'model/network/method/message/data/Checker';
 
 /** Data sent to create a Meeting event */
 export class CreateMeeting implements MessageData {
