+++ fe1-web/features/meeting/network/messages/__tests__/StateMeeting.test.ts
@@ -1,10 +1,12 @@
 import 'jest-extended';
+
 import '__tests__/utils/matchers';
 import { ActionType, ObjectType } from 'model/network/method/message/data/MessageData';
 import { Base64UrlData, Hash, Timestamp } from 'model/objects';
 import { mockLao, mockLaoId } from '__tests__/utils/TestUtils';
 import { ProtocolError } from 'model/network/ProtocolError';
 import { OpenedLaoStore } from 'store';
+
 import { StateMeeting } from '../StateMeeting';
 
 const NAME = 'myMeeting';
