+++ fe1-web/features/meeting/objects/index.ts
@@ -1,0 +1,1 @@
+export * from './Meeting';
