+++ fe1-web/model/network/method/message/data/index.ts
@@ -2,7 +2,6 @@ export * from './MessageData';
 export * from './Checker';
 export * from './witness';
 export * from './lao';
-export * from './meeting';
 export * from './rollCall';
 export * from './election';
 export * from './MessageRegistry';
