+++ fe1-web/navigation/bars/organizer/OrganizerNavigation.tsx
@@ -6,7 +6,7 @@ import STRINGS from 'res/strings';
 import Organizer from 'parts/lao/organizer/Organizer';
 import CreateEvent from 'parts/lao/organizer/eventCreation/CreateEvent';
 import WitnessScanning from 'components/WitnessScanning';
-import CreateMeeting from 'parts/lao/organizer/eventCreation/events/CreateMeeting';
+import CreateMeeting from 'features/meeting/screens/CreateMeeting';
 import CreateRollCall from 'parts/lao/organizer/eventCreation/events/CreateRollCall';
 import CreateElection from 'parts/lao/organizer/eventCreation/events/CreateElection';
 import RollCallOpened from 'parts/lao/organizer/RollCallOpened';
